requirements (6).txt
dnspython==2.3.1
prettytable==3.7.0
requests==2.31.0